+++ src/job_manager/services.py
@@ -1,6 +1,8 @@
 from datetime import datetime
 
+from api.permission_checker import AbstractPermissionService
 from common.services import model_update
+from django.core.exceptions import PermissionDenied
 from django.db import transaction
 
 from job_manager.models import (
@@ -18,11 +20,12 @@ from job_manager.models import (
 # ------------------------------------------------------------------------------
 
 
-class WorkCenterService:
-    def __init__(self, user=None):
-        self.user = user
-
+class WorkCenterService(AbstractPermissionService):
     def create(self, name: str, notes: str) -> WorkCenter:
+        # check for permission to create work center
+        if not self.check_for_permission("add_workcenter"):
+            raise PermissionDenied()
+
         work_center = WorkCenter.objects.create(name=name, notes=notes)
         work_center.full_clean()
         work_center.save(user=self.user)
@@ -30,6 +33,10 @@ class WorkCenterService:
         return work_center
 
     def update(self, work_center: WorkCenter, data: dict) -> WorkCenter:
+        # check for permission to update work center
+        if not self.check_for_permission("change_workcenter"):
+            raise PermissionDenied()
+
         fields = [
             "name",
             "notes",
@@ -42,6 +49,10 @@ class WorkCenterService:
         return work_center
 
     def delete(self, work_center: WorkCenter) -> None:
+        # check for permission to delete work center
+        if not self.check_for_permission("delete_workcenter"):
+            raise PermissionDenied()
+
         work_center.delete()
 
 
@@ -50,11 +61,12 @@ class WorkCenterService:
 # ------------------------------------------------------------------------------
 
 
-class TaskTypeService:
-    def __init__(self, user=None):
-        self.user = user
-
+class TaskTypeService(AbstractPermissionService):
     def create(self, name: str) -> TaskType:
+        # check for permission to create task type
+        if not self.check_for_permission("add_tasktype"):
+            raise PermissionDenied()
+
         task_type = TaskType.objects.create(name=name)
         task_type.full_clean()
         task_type.save(user=self.user)
@@ -62,6 +74,10 @@ class TaskTypeService:
         return task_type
 
     def update(self, task_type: TaskType, data: dict) -> TaskType:
+        # check for permission to update task type
+        if not self.check_for_permission("change_tasktype"):
+            raise PermissionDenied()
+
         fields = [
             "name",
         ]
@@ -72,6 +88,10 @@ class TaskTypeService:
         return task_type
 
     def delete(self, task_type: TaskType) -> None:
+        # check for permission to delete task type
+        if not self.check_for_permission("delete_tasktype"):
+            raise PermissionDenied()
+
         task_type.delete()
 
 
@@ -80,10 +100,7 @@ class TaskTypeService:
 # ------------------------------------------------------------------------------
 
 
-class TaskService:
-    def __init__(self, user=None):
-        self.user = user
-
+class TaskService(AbstractPermissionService):
     @transaction.atomic
     def create(
         self,
@@ -102,6 +119,10 @@ class TaskService:
         predecessors: list[Task] = None,
         successors: list[Task] = None,
     ) -> Task:
+        # check for permission to create task
+        if not self.check_for_permission("add_task"):
+            raise PermissionDenied()
+
         task = Task.objects.create(
             name=name,
             external_id=external_id,
@@ -131,6 +152,10 @@ class TaskService:
 
     @transaction.atomic
     def update(self, *, instance: Task, data: dict) -> Task:
+        # check for permission to update task
+        if not self.check_for_permission("change_task"):
+            raise PermissionDenied()
+
         fields = [
             "name",
             "external_id",
@@ -156,6 +181,10 @@ class TaskService:
 
     @transaction.atomic
     def delete(self, *, task: Task) -> None:
+        # check for permission to delete task
+        if not self.check_for_permission("delete_task"):
+            raise PermissionDenied()
+
         task.delete()
 
 
@@ -164,11 +193,12 @@ class TaskService:
 # ------------------------------------------------------------------------------
 
 
-class JobTypeService:
-    def __init__(self, user=None):
-        self.user = user
-
+class JobTypeService(AbstractPermissionService):
     def create(self, name: str, external_id: str = "", notes: str = "") -> JobType:
+        # check for permission to create job type
+        if not self.check_for_permission("add_jobtype"):
+            raise PermissionDenied()
+
         job_type = JobType.objects.create(
             name=name, external_id=external_id, notes=notes
         )
@@ -178,6 +208,10 @@ class JobTypeService:
         return job_type
 
     def update(self, job_type: JobType, data: dict) -> JobType:
+        # check for permission to update job type
+        if not self.check_for_permission("change_jobtype"):
+            raise PermissionDenied()
+
         fields = [
             "name",
             "external_id",
@@ -191,13 +225,14 @@ class JobTypeService:
         return job_type
 
     def delete(self, job_type: JobType) -> None:
+        # check for permission to delete job type
+        if not self.check_for_permission("delete_jobtype"):
+            raise PermissionDenied()
+
         job_type.delete()
 
 
-class JobService:
-    def __init__(self, user=None):
-        self.user = user
-
+class JobService(AbstractPermissionService):
     @transaction.atomic
     def create(
         self,
@@ -210,6 +245,10 @@ class JobService:
         notes: str = "",
         priority: int = None,
     ) -> Job:
+        # check for permission to create job
+        if not self.check_for_permission("add_job"):
+            raise PermissionDenied()
+
         job = Job.objects.create(
             name=name,
             due_date=due_date,
@@ -229,6 +268,10 @@ class JobService:
 
     @transaction.atomic
     def update(self, job: Job, data: dict) -> Job:
+        # check for permission to update job
+        if not self.check_for_permission("change_job"):
+            raise PermissionDenied()
+
         fields = [
             "name",
             "due_date",
@@ -250,6 +293,10 @@ class JobService:
 
     @transaction.atomic
     def delete(self, job: Job) -> None:
+        # check for permission to delete job
+        if not self.check_for_permission("delete_job"):
+            raise PermissionDenied()
+
         job.delete()
 
 
@@ -258,13 +305,14 @@ class JobService:
 # ------------------------------------------------------------------------------
 
 
-class DependencyTypeService:
-    def __init__(self, user=None):
-        self.user = user
-
+class DependencyTypeService(AbstractPermissionService):
     def create(
         self, name: str, external_id: str = "", notes: str = ""
     ) -> DependencyType:
+        # check for permission to create dependency type
+        if not self.check_for_permission("add_dependencytype"):
+            raise PermissionDenied()
+
         dependency_type = DependencyType.objects.create(
             name=name, external_id=external_id, notes=notes
         )
@@ -274,6 +322,10 @@ class DependencyTypeService:
         return dependency_type
 
     def update(self, dependency_type: DependencyType, data: dict) -> DependencyType:
+        # check for permission to update dependency type
+        if not self.check_for_permission("change_dependencytype"):
+            raise PermissionDenied()
+
         fields = [
             "name",
             "external_id",
@@ -290,13 +342,14 @@ class DependencyTypeService:
         return dependency_type
 
     def delete(self, dependency_type: DependencyType) -> None:
+        # check for permission to delete dependency type
+        if not self.check_for_permission("delete_dependencytype"):
+            raise PermissionDenied()
+
         dependency_type.delete()
 
 
-class DependencyService:
-    def __init__(self, user=None):
-        self.user = user
-
+class DependencyService(AbstractPermissionService):
     @transaction.atomic
     def create(
         self,
@@ -309,6 +362,10 @@ class DependencyService:
         tasks: list[Task] = None,
         jobs: list[Job] = None,
     ) -> Dependency:
+        # check for permission to create dependency
+        if not self.check_for_permission("add_dependency"):
+            raise PermissionDenied()
+
         dependency = Dependency.objects.create(
             name=name,
             dependency_type=dependency_type,
@@ -330,6 +387,10 @@ class DependencyService:
 
     @transaction.atomic
     def update(self, *, instance: Dependency, data: dict) -> Dependency:
+        # check for permission to update dependency
+        if not self.check_for_permission("change_dependency"):
+            raise PermissionDenied()
+
         fields = [
             "name",
             "dependency_type",
@@ -348,4 +409,8 @@ class DependencyService:
 
     @transaction.atomic
     def delete(self, instance: Dependency) -> None:
+        # check for permission to delete dependency
+        if not self.check_for_permission("delete_dependency"):
+            raise PermissionDenied()
+
         instance.delete()
