+++ src/resource_calendar/services.py
@@ -1,8 +1,9 @@
 from collections import defaultdict
 from datetime import datetime, time
 
+from api.permission_checker import AbstractPermissionService
 from common.services import model_update
-from django.core.exceptions import ValidationError
+from django.core.exceptions import PermissionDenied, ValidationError
 from django.db import transaction
 from resource_manager.models import Resource
 
@@ -14,10 +15,7 @@ from .models import (
 )
 
 
-class WeeklyShiftTemplateDetailService:
-    def __init__(self, user=None):
-        self.user = user
-
+class WeeklyShiftTemplateDetailService(AbstractPermissionService):
     def _parse_time(self, time_input: str | time) -> datetime.time:
         """
         Parse a time string (expected in %H:%M format) and return a time object.
@@ -43,6 +41,9 @@ class WeeklyShiftTemplateDetailService:
         """
         Create a WeeklyShiftTemplateDetail.
         """
+        # check for permissions for add weekly shift template detail
+        if not self.check_for_permission("add_weeklyshifttemplatedetail"):
+            raise PermissionDenied()
 
         weekly_shift_template_detail = WeeklyShiftTemplateDetail.objects.create(
             day_of_week=day_of_week,
@@ -58,15 +59,20 @@ class WeeklyShiftTemplateDetailService:
 
     @transaction.atomic
     def create_bulk(
-        *self,
+        self,
+        *,
         weekly_shift_template: WeeklyShiftTemplate,
         details: list[dict],
     ) -> None:
         """
         Create a list of WeeklyShiftTemplateDetails.
         """
+        # check for permissions for add weekly shift template detail
+        if not self.check_for_permission("add_weeklyshifttemplatedetail"):
+            raise PermissionDenied()
+
         for detail_data in details:
-            WeeklyShiftTemplateDetailService(user=self.user).create(
+            self.create(
                 day_of_week=detail_data["day_of_week"],
                 start_time=detail_data["start_time"],
                 end_time=detail_data["end_time"],
@@ -78,13 +84,14 @@ class WeeklyShiftTemplateDetailService:
         """
         Delete a WeeklyShiftTemplateDetail.
         """
+        # check for permissions for delete weekly shift template detail
+        if not self.check_for_permission("delete_weeklyshifttemplatedetail"):
+            raise PermissionDenied()
+
         instance.delete()
 
 
-class WeeklyShiftTemplateService:
-    def __init__(self, user=None):
-        self.user = user
-
+class WeeklyShiftTemplateService(AbstractPermissionService):
     def _process_details(
         self, template: WeeklyShiftTemplate, new_details: list[dict]
     ) -> None:
@@ -167,6 +174,10 @@ class WeeklyShiftTemplateService:
         """
         Create a WeeklyShiftTemplate and its related WeeklyShiftTemplateDetails.
         """
+        # check for permissions for add weekly shift template
+        if not self.check_for_permission("add_weeklyshifttemplate"):
+            raise PermissionDenied()
+
         self._validate_details_fields(details)
 
         # Create WeeklyShiftTemplate
@@ -196,6 +207,9 @@ class WeeklyShiftTemplateService:
         """
         Update a WeeklyShiftTemplate and its related WeeklyShiftTemplateDetails.
         """
+        # check for permissions for change weekly shift template
+        if not self.check_for_permission("change_weeklyshifttemplate"):
+            raise PermissionDenied()
 
         fields = [
             "name",
@@ -203,7 +217,9 @@ class WeeklyShiftTemplateService:
             "notes",
         ]
 
-        template, _ = model_update(instance=instance, fields=fields, data=data)
+        template, _ = model_update(
+            instance=instance, fields=fields, data=data, user=self.user
+        )
 
         details = data.get("details", [])
 
@@ -223,21 +239,26 @@ class WeeklyShiftTemplateService:
         return template
 
     @transaction.atomic
-    def delete(template: WeeklyShiftTemplate) -> None:
+    def delete(self, template: WeeklyShiftTemplate) -> None:
         """
         Delete a WeeklyShiftTemplate and its related WeeklyShiftTemplateDetails.
         """
+        # check for permissions for delete weekly shift template
+        if not self.check_for_permission("delete_weeklyshifttemplate"):
+            raise PermissionDenied()
+
         template.delete()
 
 
-class OperationalExceptionTypeService:
-    def __init__(self, user=None):
-        self.user = user
-
+class OperationalExceptionTypeService(AbstractPermissionService):
     @transaction.atomic
     def create(
         self, name: str, external_id: str = "", notes: str = ""
     ) -> OperationalExceptionType:
+        # check for permissions for add operational exception type
+        if not self.check_for_permission("add_operationalexceptiontype"):
+            raise PermissionDenied()
+
         exception_type = OperationalExceptionType.objects.create(
             name=name, external_id=external_id, notes=notes
         )
@@ -250,6 +271,10 @@ class OperationalExceptionTypeService:
     def update(
         self, exception_type: OperationalExceptionType, data: dict
     ) -> OperationalExceptionType:
+        # check for permissions for change operational exception type
+        if not self.check_for_permission("change_operationalexceptiontype"):
+            raise PermissionDenied()
+
         fields = [
             "name",
             "external_id",
@@ -257,20 +282,24 @@ class OperationalExceptionTypeService:
         ]
 
         exception_type, _ = model_update(
-            instance=exception_type, fields=fields, data=data, user=self.user,
+            instance=exception_type,
+            fields=fields,
+            data=data,
+            user=self.user,
         )
 
         return exception_type
 
     @transaction.atomic
-    def delete(exception_type: OperationalExceptionType) -> None:
+    def delete(self, exception_type: OperationalExceptionType) -> None:
+        # check for permissions for delete operational exception type
+        if not self.check_for_permission("delete_operationalexceptiontype"):
+            raise PermissionDenied()
+
         exception_type.delete()
 
 
-class OperationalExceptionService:
-    def __init__(self, user=None):
-        self.user = user
-
+class OperationalExceptionService(AbstractPermissionService):
     @transaction.atomic
     def create(
         *,
@@ -283,6 +312,10 @@ class OperationalExceptionService:
         external_id: str = "",
         notes="",
     ) -> OperationalException:
+        # check for permissions for add operational exception
+        if not self.check_for_permission("add_operationalexception"):
+            raise PermissionDenied()
+
         exception = OperationalException.objects.create(
             resource=resource,
             start_datetime=start_datetime,
@@ -299,7 +332,13 @@ class OperationalExceptionService:
         return exception
 
     @transaction.atomic
-    def update(self, exception: OperationalException, data: dict) -> OperationalException:
+    def update(
+        self, exception: OperationalException, data: dict
+    ) -> OperationalException:
+        # check for permissions for change operational exception
+        if not self.check_for_permission("change_operationalexception"):
+            raise PermissionDenied()
+
         fields = [
             "resource",
             "start_datetime",
@@ -310,10 +349,19 @@ class OperationalExceptionService:
             "notes",
         ]
 
-        exception, _ = model_update(instance=exception, fields=fields, data=data, user=self.user,)
+        exception, _ = model_update(
+            instance=exception,
+            fields=fields,
+            data=data,
+            user=self.user,
+        )
 
         return exception
 
     @transaction.atomic
-    def delete(exception: OperationalException) -> None:
+    def delete(self, exception: OperationalException) -> None:
+        # check for permissions for delete operational exception
+        if not self.check_for_permission("delete_operationalexception"):
+            raise PermissionDenied()
+
         exception.delete()
